+++ spec/lib/clearly/query/composer_query_spec.rb
@@ -2,11 +2,28 @@ require 'spec_helper'
 
 describe Clearly::Query::Composer do
   include_context 'shared_setup'
-  let(:product_attributes) { {name: 'plastic cup',
-                                        code: '000475PC',
-                                        brand: 'Generic',
-                                        introduced_at: '2015-01-01 00:00:00',
-                                        discontinued_at: nil}}
+  let(:product_attributes) {
+    {
+        name: 'plastic cup',
+        code: '000475PC',
+        brand: 'Generic',
+        introduced_at: '2015-01-01 00:00:00',
+        discontinued_at: nil
+    }
+  }
+
+  let(:customer_attributes) {
+    {
+        name: 'first last',
+        last_contact_at: '2015-11-09 10:00:00'
+    }
+  }
+
+  let(:order_attributes) {
+    {
+
+    }
+  }
 
   it 'finds the only product' do
     product = Product.create!(product_attributes)
@@ -43,4 +60,20 @@ describe Clearly::Query::Composer do
     expect(result_item.introduced_at).to eq(product_attributes[:introduced_at])
     expect(result_item.discontinued_at).to eq(product_attributes[:discontinued_at])
   end
+
+  it 'finds the matching order using mapped field' do
+    customer = Customer.create!(customer_attributes)
+    order_pending = Order.create!(customer: customer)
+    order_shipped = Order.create!(customer: customer, shipped_at: '2015-11-09 11:00:00')
+
+    query_hash = cleaner.do({title: {contains: 'not shipped'}})
+    query_ar = composer.query(Order, query_hash)
+
+    expect(query_ar.count).to eq(1)
+
+    result_item = query_ar.to_a[0]
+    expect(result_item.shipped_at).to eq(order_pending.shipped_at)
+    expect(result_item.customer_id).to eq(order_pending.customer_id)
+  end
+
 end