+++ spec/support/models/product.rb
@@ -57,11 +57,7 @@ class Product < ActiveRecord::Base
                     }
                 ]
             }
-        ],
-        defaults: {
-            order_by: :name,
-            direction: :asc
-        }
+        ]
     }
   end
 end